+++ raft-node/src/main/java/org/distributed/model/appendentries/LogEntry.java
@@ -4,7 +4,7 @@ package org.distributed.model.appendentries;
  * @author Oleksandr Havrylenko
  **/
 public record LogEntry(
-        int term,
+        long term,
         String command
 ) {
 }
